refactor(dynamite_runtime): split dynamite_client into separate libraries
