+++ packages/dynamite/dynamite_runtime/lib/http_client.dart
@@ -1,5 +1,9 @@
 /// The dynamite client to handle http connections.
 library http_client;
 
-export 'src/dynamite_client.dart';
+export 'src/client/authentication.dart';
+export 'src/client/client.dart';
+export 'src/client/exception.dart';
+export 'src/client/request.dart';
+export 'src/client/response.dart';
 export 'src/http_extensions.dart';
